+++ src/daf/gen.py
@@ -6,11 +6,13 @@ to automatic generation of objects.
 from typing import Optional, Union, List, Dict
 from datetime import datetime, timedelta
 from typeguard import typechecked
+from copy import deepcopy
 
 from .logging.tracing import TraceLEVELS, trace
 
 from . import misc
 from . import guild
+from . import message
 from . import client
 
 import re
@@ -33,9 +35,6 @@ GUILD_GEN_REFRESH_INT = timedelta(minutes=1) # Default interval at which to scan
 class AutoGUILD:
     """
     .. versionadded:: v2.3
-
-    TODO: Implement update method
-    TODO: Implement add_message method that will add messages to guilds based on AutoCHANNEL.
 
     Used for creating instances that will return
     guild objects.
@@ -68,11 +67,13 @@ class AutoGUILD:
         "include_pattern",
         "exclude_pattern",
         "remove_after",
+        "messages",
         "logging",
         "interval",
         "cache",
         "task",
-        "_created_at"
+        "_created_at",
+        "_deleted"
     )
 
     @typechecked
@@ -80,20 +81,19 @@ class AutoGUILD:
                  include_pattern: str,
                  exclude_pattern: Optional[str] = None,
                  remove_after: Optional[Union[timedelta, datetime]] = None,
+                 messages: Optional[List[message.BaseMESSAGE]] = [],
                  logging: Optional[bool] = False,
                  interval: Optional[timedelta] = GUILD_GEN_REFRESH_INT) -> None:
         self.include_pattern = include_pattern
         self.exclude_pattern = exclude_pattern
-        self.remove_after = remove_after  # TODO: Implement auto removal
+        self.remove_after = remove_after 
+        self.messages = messages # Uninitialized template messages list that gets copied for each found guild.
         self.logging = logging
         self.interval = interval.total_seconds() # In seconds
         self.cache: Dict[guild.GUILD] = {}
         self.task = None # Holds the Task object of _find_guilds that needs to be cleaned upon garbage collection.
+        self._deleted = False
         self._created_at = datetime.now()
-
-    def __del__(self):
-        "Called when garbage collected, cleans all the tasks"
-        self.task.cancel()
 
     @property
     def guilds(self) -> List[guild.GUILD]:
@@ -107,9 +107,21 @@ class AutoGUILD:
         """
         return self._created_at
 
+    def _delete(self):
+        """
+        Sets the internal _deleted flag to True
+        and cancels main task.
+        """
+        self._deleted = True
+        self.task.cancel()
+
     def _check_state(self) -> bool:
         """
         Checks if the object is ready to be deleted.
+
+        If the object has already been deleted, return False 
+        to prevent multiple tasks from trying to remove it multiple
+        times which would result in ValueError exceptions.
 
         Returns
         ----------
@@ -118,6 +130,9 @@ class AutoGUILD:
         False
             The object is in proper state, do not delete.
         """
+        if self._deleted:
+            return False
+
         rm_after_type = type(self.remove_after)
         return (rm_after_type is timedelta and datetime.now() - self._created_at > self.remove_after or # The difference from creation time is bigger than remove_after
                 rm_after_type is datetime and datetime.now() > self.remove_after) # The current time is larger than remove_after 
@@ -125,6 +140,24 @@ class AutoGUILD:
     async def initialize(self):
         "Initializes asynchronous elements."
         self.task = asyncio.create_task(self._find_guilds())
+    
+    async def add_message(self, message: message.BaseMESSAGE):
+        """
+        Adds a copy of the passed message to each
+        guild inside cache.
+
+        Parameters
+        -------------
+        message: message.BaseMESSAGE
+            Message to add.
+
+        Raises
+        ---------
+        Any
+            Any exception raised in :py:meth:`daf.guild.GUILD.add_message`.
+        """
+        for guild in self.cache.values():
+            await guild.add_message(deepcopy(message))
 
     async def _find_guilds(self):
         """
@@ -139,11 +172,13 @@ class AutoGUILD:
                     re.search(self.exclude_pattern, dcgld.name) is None
                 ):
                     try:
-                        new_guild = guild.GUILD(snowflake=dcgld, logging=self.logging)
+                        new_guild = guild.GUILD(snowflake=dcgld,
+                                                messages=deepcopy(self.messages),
+                                                logging=self.logging)
                         await new_guild.initialize()
                         self.cache[dcgld] = new_guild
                     except Exception as exc:
-                        trace(f"[AutoGUILD:] Unable to add new object.\nReason: {exc}", TraceLEVELS.WARNING)
+                        trace(f"[AutoGUILD:] Unable to add new object.\nReason: {exc}",TraceLEVELS.WARNING)
 
             await asyncio.sleep(self.interval)
 
@@ -159,4 +194,17 @@ class AutoGUILD:
         """
         for g in self.cache.values():
             await g._advertise(type_)
+
+    async def update(self, **kwargs):
+        """
+        Updates the object with new initialization parameters.
+
+        .. WARNING::
+            After calling this method the entire object is reset (this includes it's GUILD objects in cache).
+        """
+        if "interval" not in kwargs:
+            kwargs["interval"] = timedelta(seconds=self.interval)
+
+        return await misc._update(self, **kwargs)
+